Fiz alternancia entre páginas. coloquei mais coisas nos componentes e limpei o app.js. Arrumei função delete
